Fix possible multiplication overflows

+++ ledfeature.h
@@ -30,7 +30,7 @@ class LEDFeature : public ILEDFeature
     uint32_t    _clientBufferCount;
     shared_ptr<ISocketChannel> _ptrSocketChannel;
     static atomic<uint32_t> _nextId;
-    uint32_t _id;    
+    uint32_t _id;
 
 public:
     LEDFeature(const string & hostName,
@@ -57,9 +57,9 @@ public:
         _ptrSocketChannel = make_shared<SocketChannel>(hostName, friendlyName, port);
     }
 
-    uint32_t Id() const override 
-    { 
-        return _id; 
+    uint32_t Id() const override
+    {
+        return _id;
     }
 
     // Accessor methods
@@ -85,18 +85,18 @@ public:
         constexpr auto kBufferFillRatio = 0.80;
         return(_clientBufferCount * kBufferFillRatio) / _canvas->Effects().GetFPS();
     }
-    
-    virtual shared_ptr<ISocketChannel> Socket() override 
+
+    virtual shared_ptr<ISocketChannel> Socket() override
     {
         return _ptrSocketChannel;
     }
 
-    virtual const shared_ptr<ISocketChannel> Socket() const override 
+    virtual const shared_ptr<ISocketChannel> Socket() const override
     {
         return _ptrSocketChannel;
     }
-    
-    vector<uint8_t> GetPixelData() const override 
+
+    vector<uint8_t> GetPixelData() const override
     {
         static_assert(sizeof(CRGB) == 3, "CRGB must be 3 bytes in size for this code to work.");
 
@@ -110,8 +110,8 @@ public:
             return Utilities::ConvertPixelsToByteArray(graphics.GetPixels(), _reversed, _redGreenSwap);
 
         // Pre-calculate the final buffer size (3 bytes per pixel)
-        vector<uint8_t> result(_width * _height * sizeof(CRGB));
-        
+        vector<uint8_t> result(size_t(_width) * _height * sizeof(CRGB));
+
         // Direct byte manipulation instead of intermediate CRGB vector
         for (uint32_t y = 0; y < _height; ++y)
         {
@@ -119,10 +119,10 @@ public:
             {
                 uint32_t canvasX = x + _offsetX;
                 uint32_t canvasY = y + _offsetY;
-                
+
                 // Calculate output position directly in bytes
-                uint32_t byteIndex = (y * _width + x) * sizeof(CRGB);
-                
+                uint32_t byteIndex = (size_t(y) * _width + x) * sizeof(CRGB);
+
                 if (canvasX < graphics.Width() && canvasY < graphics.Height())
                 {
                     const CRGB& pixel = graphics.GetPixel(canvasX, canvasY);
@@ -132,14 +132,14 @@ public:
                         result[byteIndex + 1] = pixel.r;
                         result[byteIndex + 2] = pixel.b;
                     }
-                    else 
+                    else
                     {
                         result[byteIndex] = pixel.r;
                         result[byteIndex + 1] = pixel.g;
                         result[byteIndex + 2] = pixel.b;
                     }
                 }
-                else 
+                else
                 {
                     // Magenta for out of bounds (0xFF, 0x00, 0xFF)
                     result[byteIndex] = 0xFF;
@@ -184,7 +184,7 @@ public:
     }
 };
 
-inline void to_json(nlohmann::json& j, const ILEDFeature & feature) 
+inline void to_json(nlohmann::json& j, const ILEDFeature & feature)
 {
     j = {
             {"id",                feature.Id()},
@@ -212,7 +212,7 @@ inline void to_json(nlohmann::json& j, const ILEDFeature & feature)
         j["lastClientResponse"] = response;
 }
 
-inline void from_json(const nlohmann::json& j, shared_ptr<ILEDFeature> & feature) 
+inline void from_json(const nlohmann::json& j, shared_ptr<ILEDFeature> & feature)
 {
     // Use `at` for all fields since they are mandatory
     feature = make_shared<LEDFeature>(
